Create Exercise 3.2
Given N words (N>=2), find the number of pairs of words that begin with the same letter.

Input: In the first line of input you are given the number of words (N). Then, in the following N lines of input, you have the the words themselves.

Output: On output you should print the number of pairs of words that start with the same letter.